Fix style issues on iOS with backup flow Standardize uses of Switch component
