+++ packages/mobile/src/backup/BackupSocial.tsx
@@ -1,10 +1,11 @@
 import Button, { BtnTypes } from '@celo/react-components/components/Button'
+import Switch from '@celo/react-components/components/Switch'
 import colors from '@celo/react-components/styles/colors'
 import { fontStyles } from '@celo/react-components/styles/fonts'
 import { componentStyles } from '@celo/react-components/styles/styles'
 import * as React from 'react'
 import { WithNamespaces, withNamespaces } from 'react-i18next'
-import { ScrollView, StyleSheet, Switch, Text, View } from 'react-native'
+import { ScrollView, StyleSheet, Text, View } from 'react-native'
 import SafeAreaView from 'react-native-safe-area-view'
 import { connect } from 'react-redux'
 import { setSocialBackupCompleted } from 'src/account'
@@ -130,12 +131,7 @@ class BackupSocial extends React.Component<Props, State> {
           </View>
           {!socialBackupCompleted && (
             <View style={styles.confirmationSwitchContainer}>
-              <Switch
-                value={isConfirmChecked}
-                onValueChange={this.onPressConfirmSwitch}
-                trackColor={switchTrackColors}
-                thumbColor={colors.celoGreen}
-              />
+              <Switch value={isConfirmChecked} onValueChange={this.onPressConfirmSwitch} />
               <Text style={styles.confirmationSwitchLabel}>{t('socialBackup.confirmation')}</Text>
             </View>
           )}
@@ -181,11 +177,6 @@ const styles = StyleSheet.create({
     paddingRight: 5,
   },
 })
-
-const switchTrackColors = {
-  false: colors.inactive,
-  true: colors.celoGreen,
-}
 
 export default componentWithAnalytics(
   connect<StateProps, DispatchProps, {}, RootState>(
